@@ -6,10 +6,10 @@ default-members = ["crates/typos-cli"]
 resolver = "2"
 
 [workspace.package]
-license = "MIT OR Apache-2.0"
 repository = "https://github.com/crate-ci/typos"
+license = "MIT OR Apache-2.0"
 edition = "2021"
-rust-version = "1.75"  # MSRV
+rust-version = "1.80"  # MSRV
 include = [
   "build.rs",
   "src/**/*",
@@ -17,12 +17,11 @@ include = [
   "Cargo.lock",
   "LICENSE*",
   "README.md",
-  "benches/**/*",
   "examples/**/*"
 ]
 
 [workspace.lints.rust]
-rust_2018_idioms = "warn"
+rust_2018_idioms = { level = "warn", priority = -1 }
 unreachable_pub = "warn"
 unsafe_op_in_unsafe_fn = "warn"
 unused_lifetimes = "warn"
@@ -63,7 +62,6 @@ let_and_return = "allow"  # sometimes good to name what you are returning
 linkedlist = "warn"
 lossy_float_literal = "warn"
 macro_use_imports = "warn"
-match_wildcard_for_single_variants = "warn"
 mem_forget = "warn"
 mutex_integer = "warn"
 needless_for_each = "warn"
@@ -74,10 +72,10 @@ rc_mutex = "warn"
 redundant_feature_names = "warn"
 ref_option_ref = "warn"
 rest_pat_in_fully_bound_structs = "warn"
+result_large_err = "allow"
 same_functions_in_if_condition = "warn"
 self_named_module_files = "warn"
 semicolon_if_nothing_returned = "warn"
-single_match_else = "warn"
 str_to_string = "warn"
 string_add = "warn"
 string_add_assign = "warn"
@@ -85,6 +83,7 @@ string_lit_as_bytes = "warn"
 string_to_string = "warn"
 todo = "warn"
 trait_duplication_in_bounds = "warn"
+uninlined_format_args = "warn"
 verbose_file_reads = "warn"
 wildcard_imports = "warn"
 zero_sized_map_values = "warn"
@@ -95,3 +94,4 @@ panic = "abort"
 [profile.release]
 panic = "abort"
 codegen-units = 1
+lto = true
