@@ -237,7 +237,7 @@ impl FileChecker for DiffTypos {
             let stdout = std::io::stdout();
             let mut handle = stdout.lock();
             for line in diff {
-                write!(handle, "{}", line)?;
+                write!(handle, "{line}")?;
             }
         }
 
@@ -474,7 +474,7 @@ fn read_file(
             let (r, written) = encoding_rs::UTF_16LE.new_decoder_with_bom_removal().decode_to_string_without_replacement(&buffer, &mut decoded, true);
             let decoded = match r {
                 encoding_rs::DecoderResult::InputEmpty => Ok(decoded),
-                _ => Err(format!("invalid UTF-16LE encoding at byte {} in {}", written, path.display())),
+                _ => Err(format!("invalid UTF-16LE encoding at byte {written} in {}", path.display())),
             };
             let buffer = report_result(decoded, Some(path), reporter)?;
             (buffer.into_bytes(), content_type)
@@ -487,7 +487,7 @@ fn read_file(
             let (r, written) = encoding_rs::UTF_16BE.new_decoder_with_bom_removal().decode_to_string_without_replacement(&buffer, &mut decoded, true);
             let decoded = match r {
                 encoding_rs::DecoderResult::InputEmpty => Ok(decoded),
-                _ => Err(format!("invalid UTF-16BE encoding at byte {} in {}", written, path.display())),
+                _ => Err(format!("invalid UTF-16BE encoding at byte {written} in {}", path.display())),
             };
             let buffer = report_result(decoded, Some(path), reporter)?;
             (buffer.into_bytes(), content_type)
@@ -669,9 +669,10 @@ pub fn walk_path(
     checks: &dyn FileChecker,
     engine: &crate::policy::ConfigEngine<'_>,
     reporter: &dyn report::Report,
+    force_exclude: bool,
 ) -> Result<(), ignore::Error> {
     for entry in walk {
-        walk_entry(entry, checks, engine, reporter)?;
+        walk_entry(entry, checks, engine, reporter, force_exclude)?;
     }
     Ok(())
 }
@@ -681,11 +682,12 @@ pub fn walk_path_parallel(
     checks: &dyn FileChecker,
     engine: &crate::policy::ConfigEngine<'_>,
     reporter: &dyn report::Report,
+    force_exclude: bool,
 ) -> Result<(), ignore::Error> {
     let error: std::sync::Mutex<Result<(), ignore::Error>> = std::sync::Mutex::new(Ok(()));
     walk.run(|| {
         Box::new(|entry: Result<ignore::DirEntry, ignore::Error>| {
-            match walk_entry(entry, checks, engine, reporter) {
+            match walk_entry(entry, checks, engine, reporter, force_exclude) {
                 Ok(()) => ignore::WalkState::Continue,
                 Err(err) => {
                     *error.lock().unwrap() = Err(err);
@@ -703,6 +705,7 @@ fn walk_entry(
     checks: &dyn FileChecker,
     engine: &crate::policy::ConfigEngine<'_>,
     reporter: &dyn report::Report,
+    force_exclude: bool,
 ) -> Result<(), ignore::Error> {
     let entry = match entry {
         Ok(entry) => entry,
@@ -715,10 +718,14 @@ fn walk_entry(
         .iter()
         .any(|n| *n == entry.file_name())
     {
+        log::debug!(
+            "{}: skipping potential config file as it may have typos",
+            entry.path().display()
+        );
         return Ok(());
     }
     if entry.file_type().map(|t| t.is_file()).unwrap_or(true) {
-        let explicit = entry.depth() == 0;
+        let explicit = entry.depth() == 0 && !force_exclude;
         let (path, lookup_path) = if entry.is_stdin() {
             let path = std::path::Path::new("-");
             let cwd = std::env::current_dir().map_err(|err| {
