@@ -10,19 +10,24 @@
 /// columns (inclusive).
 #[rustfmt::skip]
 pub(crate) const DEFAULT_TYPES: &[(&str, &[&str])] = &[
+    ("ada", &["*.adb", "*.ads"]),
     ("agda", &["*.agda", "*.lagda"]),
     ("aidl", &["*.aidl"]),
+    ("alire", &["alire.toml"]),
     ("amake", &["*.mk", "*.bp"]),
     ("asciidoc", &["*.adoc", "*.asc", "*.asciidoc"]),
     ("asm", &["*.asm", "*.s", "*.S"]),
     ("asp", &[
-        "*.aspx", "*.aspx.cs", "*.aspx.vb", "*.ascx", "*.ascx.cs", "*.ascx.vb",
+        "*.aspx", "*.aspx.cs", "*.aspx.vb", "*.ascx", "*.ascx.cs",
+        "*.ascx.vb", "*.asp"
     ]),
     ("ats", &["*.ats", "*.dats", "*.sats", "*.hats"]),
     ("avro", &["*.avdl", "*.avpr", "*.avsc"]),
     ("awk", &["*.awk"]),
+    ("bat", &["*.bat", "*.cmd"]),
     ("bazel", &[
-        "*.bazel", "*.bzl", "*.BUILD", "*.bazelrc", "BUILD", "WORKSPACE",
+        "*.bazel", "*.bzl", "*.BUILD", "*.bazelrc", "BUILD", "MODULE.bazel",
+        "WORKSPACE", "WORKSPACE.bazel",
     ]),
     ("bitbake", &["*.bb", "*.bbappend", "*.bbclass", "*.conf", "*.inc"]),
     ("brotli", &["*.br"]),
@@ -30,6 +35,8 @@ pub(crate) const DEFAULT_TYPES: &[(&str, &[&str])] = &[
     ("bzip2", &["*.bz2", "*.tbz2"]),
     ("c", &["*.[chH]", "*.[chH].in", "*.cats"]),
     ("cabal", &["*.cabal"]),
+    ("candid", &["*.did"]),
+    ("carp", &["*.carp"]),
     ("cbor", &["*.cbor"]),
     ("cert", &[
         // Certificate files:
@@ -51,6 +58,7 @@ pub(crate) const DEFAULT_TYPES: &[(&str, &[&str])] = &[
     ("ceylon", &["*.ceylon"]),
     ("clojure", &["*.clj", "*.cljc", "*.cljs", "*.cljx"]),
     ("cmake", &["*.cmake", "CMakeLists.txt"]),
+    ("cml", &["*.cml"]),
     ("coffeescript", &["*.coffee"]),
     ("config", &["*.cfg", "*.conf", "*.config", "*.ini"]),
     ("coq", &["*.v"]),
@@ -68,14 +76,17 @@ pub(crate) const DEFAULT_TYPES: &[(&str, &[&str])] = &[
     ("cython", &["*.pyx", "*.pxi", "*.pxd"]),
     ("d", &["*.d"]),
     ("dart", &["*.dart"]),
+    ("devicetree", &["*.dts", "*.dtsi", "*.dtso"]),
     ("dhall", &["*.dhall"]),
     ("diff", &["*.patch", "*.diff"]),
+    ("dita", &["*.dita", "*.ditamap", "*.ditaval"]),
     ("docker", &["*Dockerfile*"]),
+    ("dockercompose", &["docker-compose.yml", "docker-compose.*.yml"]),
     ("dvc", &["Dvcfile", "*.dvc"]),
-    ("ebuild", &["*.ebuild"]),
+    ("ebuild", &["*.ebuild", "*.eclass"]),
     ("edn", &["*.edn"]),
     ("elisp", &["*.el"]),
-    ("elixir", &["*.ex", "*.eex", "*.exs"]),
+    ("elixir", &["*.ex", "*.eex", "*.exs", "*.heex", "*.leex", "*.livemd"]),
     ("elm", &["*.elm"]),
     ("erb", &["*.erb"]),
     ("erlang", &["*.erl", "*.hrl"]),
@@ -92,13 +103,19 @@ pub(crate) const DEFAULT_TYPES: &[(&str, &[&str])] = &[
     ("gap", &["*.g", "*.gap", "*.gi", "*.gd", "*.tst"]),
     ("gn", &["*.gn", "*.gni"]),
     ("go", &["*.go"]),
-    ("gradle", &["*.gradle"]),
+    ("gprbuild", &["*.gpr"]),
+    ("gradle", &[
+        "*.gradle", "*.gradle.kts", "gradle.properties", "gradle-wrapper.*",
+        "gradlew", "gradlew.bat",
+    ]),
+    ("graphql", &["*.graphql", "*.graphqls"]),
     ("groovy", &["*.groovy", "*.gradle"]),
     ("gzip", &["*.gz", "*.tgz"]),
     ("h", &["*.h", "*.hpp"]),
     ("haml", &["*.haml"]),
+    ("hare", &["*.ha"]),
     ("hbs", &["*.hbs"]),
-    ("hs", &["*.hs", "*.lhs"]),
+    ("hs", &["*.hs", "*.lhs", "*.cpphs", "*.c2hs", "*.hsc"]),
     ("html", &["*.htm", "*.html", "*.ejs"]),
     ("hy", &["*.hy"]),
     ("idris", &["*.idr", "*.lidr"]),
@@ -106,12 +123,13 @@ pub(crate) const DEFAULT_TYPES: &[(&str, &[&str])] = &[
     ("java", &["*.java", "*.jsp", "*.jspx", "*.properties"]),
     ("jinja", &["*.j2", "*.jinja", "*.jinja2"]),
     ("jl", &["*.jl"]),
-    ("js", &["*.js", "*.jsx", "*.vue"]),
+    ("js", &["*.js", "*.jsx", "*.vue", "*.cjs", "*.mjs"]),
     ("json", &["*.json"]),
     ("jsonl", &["*.jsonl"]),
     ("jupyter", &["*.ipynb", "*.jpynb"]),
     ("k", &["*.k"]),
     ("kotlin", &["*.kt", "*.kts"]),
+    ("lean", &["*.lean"]),
     ("less", &["*.less"]),
     ("license", &[
         // General
@@ -142,7 +160,13 @@ pub(crate) const DEFAULT_TYPES: &[(&str, &[&str])] = &[
     ]),
     ("lilypond", &["*.ly", "*.ily"]),
     ("lisp", &["*.el", "*.lisp", "*.lsp", "*.sc", "*.scm"]),
-    ("lock", &["*.lock", "package-lock.json", "requirements.txt", "go.sum", "pnpm-lock.yaml"]),
+    ("lock", &[
+        "*.lock",
+        "package-lock.json", "pnpm-lock.yaml",
+        "composer.lock", "*.terraform.lock.hcl",
+        "requirements.txt",
+        "go.sum",
+    ]),
     ("log", &["*.log"]),
     ("lua", &["*.lua"]),
     ("lz4", &["*.lz4"]),
@@ -157,14 +181,24 @@ pub(crate) const DEFAULT_TYPES: &[(&str, &[&str])] = &[
     ("mako", &["*.mako", "*.mao"]),
     ("man", &["*.[0-9lnpx]", "*.[0-9][cEFMmpSx]"]),
     ("matlab", &["*.m"]),
-    ("md", &["*.markdown", "*.md", "*.mdown", "*.mkdn"]),
-    ("meson", &["meson.build", "meson_options.txt"]),
+    ("md", &[
+        "*.markdown",
+        "*.md",
+        "*.mdown",
+        "*.mdwn",
+        "*.mkd",
+        "*.mkdn",
+        "*.mdx",
+    ]),
+    ("meson", &["meson.build", "meson_options.txt", "meson.options"]),
     ("minified", &["*.min.html", "*.min.css", "*.min.js"]),
     ("mint", &["*.mint"]),
     ("mk", &["mkfile"]),
     ("ml", &["*.ml"]),
+    ("motoko", &["*.mo"]),
     ("msbuild", &[
         "*.csproj", "*.fsproj", "*.vcxproj", "*.proj", "*.props", "*.targets",
+        "*.sln",
     ]),
     ("nim", &["*.nim", "*.nimf", "*.nimble", "*.nims"]),
     ("nix", &["*.nix"]),
@@ -175,13 +209,19 @@ pub(crate) const DEFAULT_TYPES: &[(&str, &[&str])] = &[
     ("pascal", &["*.pas", "*.dpr", "*.lpr", "*.pp", "*.inc"]),
     ("pdf", &["*.pdf"]),
     ("perl", &["*.perl", "*.pl", "*.PL", "*.plh", "*.plx", "*.pm", "*.t"]),
-    ("php", &["*.php", "*.php3", "*.php4", "*.php5", "*.phtml"]),
+    ("php", &[
+        // note that PHP 6 doesn't exist
+        // See: https://wiki.php.net/rfc/php6
+        "*.php", "*.php3", "*.php4", "*.php5", "*.php7", "*.php8",
+        "*.pht", "*.phtml"
+    ]),
     ("po", &["*.po"]),
     ("pod", &["*.pod"]),
     ("postscript", &["*.eps", "*.ps"]),
+    ("prolog", &["*.prolog", "*.P"]),
     ("protobuf", &["*.proto"]),
     ("ps", &["*.cdxml", "*.ps1", "*.ps1xml", "*.psd1", "*.psm1"]),
-    ("puppet", &["*.erb", "*.pp", "*.rb"]),
+    ("puppet", &["*.epp", "*.erb", "*.pp", "*.rb"]),
     ("purs", &["*.purs"]),
     ("py", &[
         "*.py",
@@ -193,9 +233,15 @@ pub(crate) const DEFAULT_TYPES: &[(&str, &[&str])] = &[
     ("qml", &["*.qml"]),
     ("r", &["*.R", "*.r", "*.Rmd", "*.Rnw"]),
     ("racket", &["*.rkt"]),
+    ("raku", &[
+        "*.raku", "*.rakumod", "*.rakudoc", "*.rakutest",
+        "*.p6", "*.pl6", "*.pm6"
+    ]),
     ("rdoc", &["*.rdoc"]),
     ("readme", &["README*", "*README"]),
+    ("reasonml", &["*.re", "*.rei"]),
     ("red", &["*.r", "*.red", "*.reds"]),
+    ("rescript", &["*.res", "*.resi"]),
     ("robot", &["*.robot"]),
     ("rst", &["*.rst"]),
     ("ruby", &[
@@ -237,12 +283,14 @@ pub(crate) const DEFAULT_TYPES: &[(&str, &[&str])] = &[
     ("slim", &["*.skim", "*.slim", "*.slime"]),
     ("smarty", &["*.tpl"]),
     ("sml", &["*.sml", "*.sig"]),
+    ("solidity", &["*.sol"]),
     ("soy", &["*.soy"]),
     ("spark", &["*.spark"]),
     ("spec", &["*.spec"]),
     ("sql", &["*.sql", "*.psql"]),
     ("stylus", &["*.styl"]),
     ("sv", &["*.v", "*.vg", "*.sv", "*.svh", "*.h"]),
+    ("svelte", &["*.svelte"]),
     ("svg", &["*.svg"]),
     ("swift", &["*.swift"]),
     ("swig", &["*.def", "*.i"]),
@@ -256,13 +304,19 @@ pub(crate) const DEFAULT_TYPES: &[(&str, &[&str])] = &[
     ("tex", &["*.tex", "*.ltx", "*.cls", "*.sty", "*.bib", "*.dtx", "*.ins"]),
     ("texinfo", &["*.texi"]),
     ("textile", &["*.textile"]),
-    ("tf", &["*.tf"]),
+    ("tf", &[
+        "*.tf", "*.auto.tfvars", "terraform.tfvars", "*.tf.json",
+        "*.auto.tfvars.json", "terraform.tfvars.json", "*.terraformrc",
+        "terraform.rc", "*.tfrc",
+    ]),
     ("thrift", &["*.thrift"]),
     ("toml", &["*.toml"]),
-    ("ts", &["*.ts", "*.tsx"]),
+    ("ts", &["*.ts", "*.tsx", "*.cts", "*.mts"]),
     ("twig", &["*.twig"]),
     ("txt", &["*.txt"]),
     ("typoscript", &["*.typoscript"]),
+    ("usd", &["*.usd", "*.usda", "*.usdc"]),
+    ("v", &["*.v", "*.vsh"]),
     ("vala", &["*.vala"]),
     ("vb", &["*.vb"]),
     ("vcl", &["*.vcl"]),
@@ -272,6 +326,7 @@ pub(crate) const DEFAULT_TYPES: &[(&str, &[&str])] = &[
         "*.vim", ".vimrc", ".gvimrc", "vimrc", "gvimrc", "_vimrc", "_gvimrc",
     ]),
     ("webidl", &["*.idl", "*.webidl", "*.widl"]),
+    ("wgsl", &["*.wgsl"]),
     ("wiki", &["*.mediawiki", "*.wiki"]),
     ("xml", &[
         "*.xml", "*.xml.dist", "*.dtd", "*.xsl", "*.xslt", "*.xsd", "*.xjb",
