@@ -8,10 +8,28 @@ pub(crate) const NO_CHECK_TYPES: &[&str] = &["cert", "lock"];
 
 pub(crate) const TYPE_SPECIFIC_DICTS: &[(&str, StaticDictConfig)] = &[
     (
+        "cpp",
+        StaticDictConfig {
+            ignore_idents: &[
+                "countr_one", // `std::countr_one`
+            ],
+            ignore_words: &[],
+        },
+    ),
+    (
         "css",
         StaticDictConfig {
             ignore_idents: &[
                 "nd", // CSS class used by pygments (see https://github.com/pygments/pygments/blob/2.16.1/pygments/token.py#L146)
+            ],
+            ignore_words: &[],
+        },
+    ),
+    (
+        "go",
+        StaticDictConfig {
+            ignore_idents: &[
+                "flate", // https://pkg.go.dev/compress/flate
             ],
             ignore_words: &[],
         },
@@ -88,6 +106,7 @@ pub(crate) struct StaticDictConfig {
 #[cfg(test)]
 mod tests {
     use itertools::Itertools;
+    use snapbox::prelude::*;
 
     use super::TYPE_SPECIFIC_DICTS;
 
@@ -96,7 +115,7 @@ mod tests {
         let types: Vec<_> = TYPE_SPECIFIC_DICTS.iter().map(|(typ, _)| *typ).collect();
         let types_unique: Vec<_> = types.clone().into_iter().unique().collect();
 
-        snapbox::assert_eq(types.join("\n"), types_unique.join("\n"));
+        snapbox::assert_data_eq!(types_unique.join("\n"), types.join("\n").raw());
     }
 
     #[test]
@@ -108,6 +127,6 @@ mod tests {
         let types: Vec<_> = TYPE_SPECIFIC_DICTS.iter().map(|(typ, _)| *typ).collect();
         let types_sorted: Vec<_> = types.iter().cloned().sorted().collect();
 
-        snapbox::assert_eq(types.join("\n"), types_sorted.join("\n"));
+        snapbox::assert_data_eq!(types_sorted.join("\n"), types.join("\n").raw());
     }
 }
