@@ -227,13 +227,16 @@ impl<'s> ConfigEngine<'s> {
         let check_filename = engine.check_filename();
         let check_file = engine.check_file();
         let crate::config::EngineConfig {
-            tokenizer: mut tokenizer_config,
-            dict: mut dict_config,
+            tokenizer: tokenizer_user_config,
+            dict: dict_user_config,
             extend_ignore_re,
             ..
         } = engine;
-        tokenizer_config.update(&crate::config::TokenizerConfig::from_defaults());
-        dict_config.update(&crate::config::DictConfig::from_defaults());
+
+        let mut tokenizer_config = crate::config::TokenizerConfig::from_defaults();
+        tokenizer_config.update(&tokenizer_user_config);
+        let mut dict_config = crate::config::DictConfig::from_defaults();
+        dict_config.update(&dict_user_config);
 
         if !tokenizer_config.ignore_hex() {
             log::warn!("`ignore-hex` is deprecated");
@@ -319,7 +322,7 @@ impl DirConfig {
 
         let config = name
             .and_then(|name| {
-                log::debug!("{}: `{}` policy", path.display(), name);
+                log::debug!("{}: `{name}` policy", path.display());
                 self.types.get(name).copied()
             })
             .unwrap_or_else(|| {
@@ -356,7 +359,7 @@ pub struct Policy<'t, 'd, 'i> {
     pub ignore: &'i [regex::Regex],
 }
 
-impl<'t, 'd, 'i> Policy<'t, 'd, 'i> {
+impl Policy<'_, '_, '_> {
     pub fn new() -> Self {
         Default::default()
     }
@@ -366,7 +369,7 @@ static DEFAULT_TOKENIZER: typos::tokens::Tokenizer = typos::tokens::Tokenizer::n
 static DEFAULT_DICT: crate::dict::BuiltIn = crate::dict::BuiltIn::new(crate::config::Locale::En);
 static DEFAULT_IGNORE: &[regex::Regex] = &[];
 
-impl<'t, 'd, 'i> Default for Policy<'t, 'd, 'i> {
+impl Default for Policy<'_, '_, '_> {
     fn default() -> Self {
         Self {
             check_filenames: true,
