@@ -4,4 +4,4 @@
 
 mod word_codegen;
 
-pub use crate::word_codegen::WORD_TRIE;
+pub use crate::word_codegen::WORD;
