@@ -1,6 +1,8 @@
 #[derive(Copy, Clone, PartialEq, Eq, Hash, Debug)]
 pub struct Cluster {
-    pub header: Option<&'static str>,
+    pub header: &'static str,
+    pub verified: bool,
+    pub level: usize,
     pub entries: &'static [Entry],
     pub notes: &'static [&'static str],
 }
@@ -8,7 +10,9 @@ pub struct Cluster {
 impl Cluster {
     pub fn into_owned(self) -> crate::Cluster {
         crate::Cluster {
-            header: self.header.map(|s| s.to_owned()),
+            header: self.header.to_owned(),
+            verified: self.verified,
+            level: self.level,
             entries: self.entries.iter().map(|s| s.into_owned()).collect(),
             notes: self.notes.iter().map(|s| (*s).to_owned()).collect(),
         }
@@ -20,8 +24,8 @@ pub struct Entry {
     pub variants: &'static [Variant],
     pub pos: Option<crate::Pos>,
     pub archaic: bool,
-    pub note: bool,
     pub description: Option<&'static str>,
+    pub note: Option<&'static str>,
     pub comment: Option<&'static str>,
 }
 
@@ -31,8 +35,8 @@ impl Entry {
             variants: self.variants.iter().map(|v| v.into_owned()).collect(),
             pos: self.pos,
             archaic: self.archaic,
-            note: self.note,
             description: self.description.map(|s| s.to_owned()),
+            note: self.note.map(|s| s.to_owned()),
             comment: self.comment.map(|s| s.to_owned()),
         }
     }
